1) Updated the Course Entity, and associated service, validator, and repository. 2) Migrated existing Course entity, service, and repository over to CourseJoin. This allowed me to create the proper Course Entity, service, and repository. 3) Cleaned up the existing services, requests, and repositories to reflect these changes 4) Ensured that all Id's that are referenced are configured as Long (BigINT). This required some database changes to get it consistent. 5) Updated the TestWidow Entity to ensure Id's were Long, and that references to Requests, services, and Entities were reflected after refactoring.

+++ backend/src/main/java/org/mentats/mentat/components/ExamResultValidator.java
@@ -12,8 +12,7 @@ import java.util.Date;
  */
 @Component
 public class ExamResultValidator {
-
-    // Existing methods...
+    // Exam Result Validator functions
     public void validateForCreation(ExamResult examResult) {
         validateNotNull(examResult, "Exam result");
         validateRequiredFields(examResult);
@@ -22,19 +21,19 @@ public class ExamResultValidator {
     }
 
     // NEW METHOD: Validate exam result ID
-    public void validateExamResultId(Integer id) {
+    public void validateExamResultId(Long id) {
         if (id == null || id <= 0) {
             throw new ValidationException("Exam result ID must be a positive number");
         }
     }
 
-    public void validateExamId(Integer examId) {
+    public void validateExamId(Long examId) {
         if (examId == null || examId <= 0) {
             throw new ValidationException("Exam ID must be a positive number");
         }
     }
 
-    public void validateStudentId(Integer studentId) {
+    public void validateStudentId(Long studentId) {
         if (studentId == null || studentId <= 0) {
             throw new ValidationException("Student ID must be a positive number");
         }
